requirements.txt
fastapi
uvicorn[standard]
pydantic
chromadb
sentence-transformers
python-dotenv
langchain
langchain-core
langchain-openai
python-dotenv

